DPDK: IPv6 support
Also change "dpdk" arg by "--dpdk" which is more standard
Do not hardcode the SERVER_ADDRESS anymore, it can be given with "-d",
both IPv4 and IPv6

Change the way tx_buffer are flushed. We count the number of sent
packets, and we call for flush after sending all we can send, or if
there's congestion.

TODO: The tx-buffer should be allocated
on the stack, no need to pass it to the sockloop